dev to master

See merge request ibs-repo/csc309-winter-2023/group_647/p3!18

+++ frontend/src/components/Logout.jsx
@@ -1,9 +1,15 @@
-import { useNavigate } from 'react-router-dom'
-import axios from 'axios';
+import React, { useEffect } from "react";
+import { useNavigate } from "react-router-dom";
+import axios from "axios";
 
 export default function Logout() {
-    const navigate = useNavigate();
-    localStorage.clear();
-    axios.defaults.headers.common['Authorization'] = null;
-    return navigate('/login')
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    if (localStorage.getItem("access_token")) {
+      localStorage.clear();
+      delete axios.defaults.headers.common["Authorization"];
+    }
+    navigate("/");
+  }, []);
 }
